dashboard/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
altair>=5.0.0
duckdb>=1.3.0
boto3>=1.34.0
python-dotenv>=1.0.0
